content/templates/cookbook/image_filters/requirements.txt
ttnn
torch
numpy
matplotlib
Pillow
opencv-python
